Align the allocate logic flow with the spec flow Update the test script so that items that are covered by Constraint checking are handled as such Avoid Null Pointer exception

+++ examples/camera-app/linux/src/clusters/push-av-stream-transport/push-av-stream-manager.cpp
@@ -342,7 +342,7 @@ Protocols::InteractionModel::Status PushAvStreamTransportManager::ValidateZoneId
     return Status::Failure;
 }
 
-bool PushAvStreamTransportManager::ValidateMotionZoneSize(uint16_t zoneSize)
+bool PushAvStreamTransportManager::ValidateMotionZoneSize(size_t zoneSize)
 {
     if (mCameraDevice == nullptr)
     {
